requirements.txt
# E-T-A RAG Prototype - Requirements
# All lightweight, runs on MacBook CPU

# LLM & Embeddings
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0

# Vector Store
chromadb>=0.5.0

# Document Processing
openpyxl>=3.1.0
pandas>=2.0.0
pymupdf>=1.24.0
python-docx>=1.1.0

# BM25 Search
rank-bm25>=0.2.2

# API & UI
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.40.0

# Ollama integration
ollama>=0.4.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
